+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Explainer for trait `buffer-file-name` now mentions the actual
   issue.
+- Calling `pop-to-buffer` with NORECORD flag is now honored.
 
 ## [v0.8.3]